Abstract out generate content logic, handle errors gracefully with AI generated message
